Codes/Python/requirements.txt
absl-py==2.1.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
daal==2025.0.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
flatbuffers==24.3.25
fonttools==4.55.0
fqdn==1.5.1
gast==0.6.0
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
grpcio==1.68.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-mathjax==0.2.6
jupyter-server-terminals==0.5.3
jupyterlab==4.3.2
jupyterlab-git==0.50.2
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
marimo==0.9.32
markdown==3.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.15.2
nbclient==0.10.1
nbconvert==7.16.4
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numpy==2.0.2
opt-einsum==3.4.0
optree==0.13.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip==24.3.1
platformdirs==4.3.6
prometheus-client==0.21.1
prompt-toolkit==3.0.48
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydot==3.0.3
pygments==2.18.0
pymdown-extensions==10.12
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
send2trash==1.8.3
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tbb==2022.0.0
tcmlib==1.2.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-cpu==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
